Moved header button struct to its own file
